requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Mail==0.9.1
WTForms==3.1.0
email-validator==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # for PostgreSQL
celery==5.3.4
redis==5.0.1
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
requests==2.31.0
openpyxl==3.1.2
pandas==2.1.4
gunicorn==21.2.0